Separate mvfs into its own crate
