+++ mvfs/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "mvfs"
+version = "0.1.13"
+edition = "2021"
+license = "Apache-2.0"
+authors = ["Heyang Zhou <heyang.zhou@icloud.com>"]
+description = "mvfs"
+repository = "https://github.com/losfair/mvsqlite"
+
+[lib]
+crate-type = ["rlib", "staticlib"]
+
+[dependencies]
+anyhow = "1"
+thiserror = "1"
+tokio = { version = "1", features = ["full"] }
+mvclient = { path = "../mvclient" }
+tracing = "0.1"
+moka = "0.9.2"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+lazy_static = "1.4.0"
+reqwest = { version = "0.11.11", default-features = false } # the dependent crate should set its own features
