+++ mvfs/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod types;
+pub mod vfs;
+pub use vfs::{Connection, MultiVersionVfs};
