+++ mvfs/template_32k.db
