+++ mvsqlite/Cargo.toml
@@ -15,22 +15,19 @@ anyhow = "1"
 thiserror = "1"
 tokio = { version = "1", features = ["full"] }
 log = "0.4"
-time = "0.3"
 rand = "0.8.5"
 stackful = "0.1.5"
-async-trait = "0.1.56"
 mvclient = { path = "../mvclient" }
 tracing = "0.1"
 ctor = "0.1.22"
 tracing-subscriber = { version = "0.3.15", features = ["env-filter", "fmt", "json"], optional = true }
 libc = "0.2"
 backtrace = "0.3.66"
-moka = "0.9.2"
 serde = { version = "1", features = ["derive"] }
 serde_json = "1"
 lazy_static = "1.4.0"
 reqwest = { version = "0.11.11", default-features = false }
-base64 = "0.13.0"
+mvfs = { path = "../mvfs" }
 
 [features]
 default = ["loadext", "syscall", "default-tls", "global-init"]
